fix google calendar

+++ app/controllers/client/tasks_controller.rb
@@ -10,31 +10,38 @@ class Client::TasksController < AdminController
       summary    = "#{domain.name} - #{task.name}"
       start_date = task.completion_date.to_s
       end_date   = (task.completion_date + 1).to_s
-      primary_calendar_id = ''
-      events = ''
       client = Signet::OAuth2::Client.new(client_id: Rails.application.secrets.google_client_id,
                                           client_secret: Rails.application.secrets.google_client_secret,
                                           token_credential_uri: 'https://accounts.google.com/o/oauth2/token')
       client.update!(session[:authorization])
       service = Google::Apis::CalendarV3::CalendarService.new
       service.authorization = client
-      service.list_events('primary').items.each do |event|
-        if event.summary == summary
-          events << event.summary
-          break
-        end
-      end
-      if events.present?
-        session[:task_id] = ''
-        redirect_to url_for, alert: t('has_been_added_to_calendar')
-      else
+      event = Google::Apis::CalendarV3::Event.new(start: Google::Apis::CalendarV3::EventDateTime.new(date: start_date),
+                                                  end: Google::Apis::CalendarV3::EventDateTime.new(date: end_date),
+                                                  summary: summary)
+      service.insert_event('primary', event)
+      session[:task_id] = nil
+      flash[:notice] = t('add_event_success')
+    elsif session[:action].present?
+      tasks = Task.incomplete.of_user(current_user).incomplete.upcoming
+      client = Signet::OAuth2::Client.new(client_id: Rails.application.secrets.google_client_id,
+                                          client_secret: Rails.application.secrets.google_client_secret,
+                                          token_credential_uri: 'https://accounts.google.com/o/oauth2/token')
+      client.update!(session[:authorization])
+      service = Google::Apis::CalendarV3::CalendarService.new
+      service.authorization = client
+      tasks.each do |task|
+        domain    = Domain.find(task.domain_id)
+        summary    = "#{domain.name} - #{task.name}"
+        start_date = task.completion_date.to_s
+        end_date   = (task.completion_date + 1).to_s
         event = Google::Apis::CalendarV3::Event.new(start: Google::Apis::CalendarV3::EventDateTime.new(date: start_date),
                                                     end: Google::Apis::CalendarV3::EventDateTime.new(date: end_date),
                                                     summary: summary)
         service.insert_event('primary', event)
-        session[:task_id] = ''
-        flash[:notice] = t('add_event_success')
       end
+      session[:action] = nil
+      flash[:notice] = t('add_event_success')
     end
   end
 
@@ -106,28 +113,36 @@ class Client::TasksController < AdminController
 
   def destroy
     @task = @client.tasks.find(params[:id])
-    respond_to do |format|
-      if @task.destroy
-        task_name       = @task.name
-        domain_name     = Domain.find(@task.domain_id).name
-        completion_date = @task.completion_date.to_s
-        summary    = "#{domain_name} - #{task_name}"
-        client = Signet::OAuth2::Client.new(client_id: '304280597205-m9d4o2bcnkrjf6gr7p6n1khufdskv4kv.apps.googleusercontent.com',
-                                            client_secret: '0pdcBK3xm_RlcVvUHYTZ2NVs',
-                                            token_credential_uri: 'https://accounts.google.com/o/oauth2/token')
-        client.update!(session[:authorization])
-        service = Google::Apis::CalendarV3::CalendarService.new
-        service.authorization = client
-        service.list_events('primary').items.each do |event|
-          event_start_date = event.start.date || event.start.date_time
-          if event.summary == summary && event_start_date == completion_date
-            service.delete_event('primary', event.id)
-            break
+    if @task.destroy
+      if current_user.calendar_integration?
+        if session[:authorization].blank? || current_user.expires_at < DateTime.now.in_time_zone
+          session[:referrer] = request.referrer
+          session[:action] = params[:action]
+          session[:task_id] = @task.id
+          redirect_to redirect_path
+        else
+          task_name       = @task.name
+          domain_name     = Domain.find(@task.domain_id).name
+          completion_date = @task.completion_date.to_s
+          summary    = "#{domain_name} - #{task_name}"
+          client = Signet::OAuth2::Client.new(client_id: '304280597205-m9d4o2bcnkrjf6gr7p6n1khufdskv4kv.apps.googleusercontent.com',
+                                              client_secret: '0pdcBK3xm_RlcVvUHYTZ2NVs',
+                                              token_credential_uri: 'https://accounts.google.com/o/oauth2/token')
+          client.update!(session[:authorization])
+          service = Google::Apis::CalendarV3::CalendarService.new
+          service.authorization = client
+          service.list_events('primary').items.each do |event|
+            event_start_date = event.start.date || event.start.date_time
+            if event.summary == summary && event_start_date == completion_date
+              service.delete_event('primary', event.id)
+              break
+            end
           end
         end
+      else
+        format.json { head :ok }
+        format.html { redirect_to client_tasks_path(@client), notice: t('.successfully_deleted') }
       end
-      format.json { head :ok }
-      format.html { redirect_to client_tasks_path(@client), notice: t('.successfully_deleted') }
     end
   end
